requirements.txt
python-telegram-bot==21.4
pandas==2.2.2
scikit-learn==1.5.2
PyYAML==6.0.2
python-dotenv==1.0.1
jinja2==3.1.4
